feat: Validator 클래스 분리

+++ src/main/java/com/scheduleprojectdevelop/service/CommentService.java
@@ -1,6 +1,7 @@
 package com.scheduleprojectdevelop.service;
 
 
+import com.scheduleprojectdevelop.AuthValidator;
 import com.scheduleprojectdevelop.dto.commentDto.*;
 import com.scheduleprojectdevelop.entity.Comment;
 import com.scheduleprojectdevelop.entity.Schedule;
@@ -17,17 +18,18 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class CommentService {
-    private CommentRepository commentRepository;
-    private UserRepository userRepository;
-    private ScheduleRepository scheduleRepository;
+    private final CommentRepository commentRepository;
+    private final UserRepository userRepository;
+    private final ScheduleRepository scheduleRepository;
+    private final AuthValidator authValidator;
 
     @Transactional
     public CreateCommentResponse create(Long scheduleId, Long userId, CreateCommentRequest request) {
         User user = userRepository.findById(userId).orElseThrow(
-                () -> new UserNotFoundException("유저를 찾을 수 없습니다.")
+                UserNotFoundException::new
         );
         Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(
-                () -> new ScheduleNotFoundException("일정을 찾을 수 없습니다.")
+                ScheduleNotFoundException::new
         );
 
         Comment comment = new Comment(user, schedule, request.getComments());
@@ -37,9 +39,7 @@ public class CommentService {
 
     @Transactional(readOnly = true)
     public GetOneCommentResponse getOneComment(Long scheduleId, Long commentId) {
-        Comment comment = commentRepository.findBySchedule_ScheduleIdAndCommentId(scheduleId, commentId).orElseThrow(
-                () -> new CommentNotFoundException("댓글을 찾을 수 없습니다.")
-        );
+        Comment comment = getComment(scheduleId, commentId);
         return new GetOneCommentResponse(comment.getComments(),
                 comment.getCommentId(),
                 comment.getUser().getUserName(),
@@ -49,6 +49,7 @@ public class CommentService {
 
     @Transactional(readOnly = true)
     public List<GetOneCommentResponse> getAllComment(Long scheduleId) {
+        scheduleRepository.findById(scheduleId).orElseThrow(ScheduleNotFoundException::new);
         List<Comment> comments = commentRepository.findAllBySchedule_ScheduleId(scheduleId);
         List<GetOneCommentResponse> results = new ArrayList<>();
         for (Comment comment : comments) {
@@ -64,19 +65,26 @@ public class CommentService {
     }
 
     @Transactional
-    public UpdateCommentResponse updateComment(Long scheduleId, Long commentId, UpdateCommentRequest request) {
-        Comment comment = commentRepository.findBySchedule_ScheduleIdAndCommentId(scheduleId, commentId).orElseThrow(
-                () -> new CommentNotFoundException("댓글을 찾을 수 없습니다.")
-        );
+    public UpdateCommentResponse updateComment(Long scheduleId, Long commentId, Long userId, UpdateCommentRequest request) {
+        Comment comment = getComment(scheduleId, commentId);
+        authValidator.validateAuthor(comment.getUser().getUserId(), userId);
+
         comment.update(request.getComments());
         return new UpdateCommentResponse(comment.getComments(), comment.getCommentId(), comment.getUser().getUserName(), comment.getCreatedAt(), comment.getUpdatedAt());
     }
 
     @Transactional
-    public void deleteComment(Long scheduleId, Long commentId) {
-        Comment comment = commentRepository.findBySchedule_ScheduleIdAndCommentId(scheduleId, commentId).orElseThrow(
-                () -> new CommentNotFoundException("댓글을 찾을 수 없습니다.")
-        );
+    public void deleteComment(Long scheduleId, Long commentId, Long userId) {
+        Comment comment = getComment(scheduleId, commentId);
+        authValidator.validateAuthor(comment.getUser().getUserId(), userId);
+
         commentRepository.delete(comment);
     }
+
+    private Comment getComment(Long scheduleId, Long commentId) {
+        return commentRepository.findBySchedule_ScheduleIdAndCommentId(scheduleId, commentId).orElseThrow(
+                CommentNotFoundException::new
+        );
+    }
+
 }
