+++ src/main/java/com/scheduleprojectdevelop/service/ScheduleService.java
@@ -1,10 +1,12 @@
 package com.scheduleprojectdevelop.service;
 
+import com.scheduleprojectdevelop.AuthValidator;
 import com.scheduleprojectdevelop.dto.scheduleDto.*;
 import com.scheduleprojectdevelop.entity.Schedule;
 import com.scheduleprojectdevelop.entity.User;
 import com.scheduleprojectdevelop.exception.ArgumentMismatchException;
 import com.scheduleprojectdevelop.exception.ScheduleNotFoundException;
+import com.scheduleprojectdevelop.exception.UserNotFoundException;
 import com.scheduleprojectdevelop.repository.ScheduleRepository;
 import com.scheduleprojectdevelop.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,19 +17,17 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.List;
-
 @Service
 @RequiredArgsConstructor
 public class ScheduleService {
     private final ScheduleRepository scheduleRepository;
     private final UserRepository userRepository;
+    private final AuthValidator authValidator;
 
     @Transactional
     public CreateScheduleResponse createSchedule(CreateScheduleRequest request, Long userId) {
         User user = userRepository.findById(userId).orElseThrow(
-                () -> new ArgumentMismatchException("어쩌고~")
+                UserNotFoundException::new
         );
         Schedule schedule = new Schedule(request.getTitle(), request.getContents(), user);
         scheduleRepository.save(schedule);
@@ -36,17 +36,14 @@ public class ScheduleService {
                 schedule.getScheduleId(),
                 schedule.getTitle(),
                 schedule.getContent(),
-                schedule.getUser(),
+                schedule.getUser().getUserId(),
                 schedule.getCreatedAt(),
                 schedule.getUpdatedAt());
     }
 
     @Transactional(readOnly = true)
     public GetOneScheduleResponse getOneSchedule(Long scheduleId) {
-        Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(
-                () -> new ScheduleNotFoundException("존재하지 않는 일정입니다.")
-        );
-
+        Schedule schedule = getSchedule(scheduleId);
 
         return new GetOneScheduleResponse(
                 schedule.getScheduleId(),
@@ -71,9 +68,8 @@ public class ScheduleService {
 
     @Transactional
     public UpdateScheduleResponse updateSchedule(Long scheduleId, UpdateScheduleRequest request, Long userId) {
-        Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(
-                () -> new ScheduleNotFoundException("존재하지 않는 일정입니다.")
-        );
+        Schedule schedule = getSchedule(scheduleId);
+        authValidator.validateAuthor(schedule.getUser().getUserId(), userId);
 
         schedule.update(request.getTitle(), request.getContents());
 
@@ -88,12 +84,17 @@ public class ScheduleService {
     }
 
     @Transactional
-    public void deleteSchedule(Long scheduleId) {
-        Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(
-                () -> new ScheduleNotFoundException("존재하지 않는 일정입니다.")
-        );
+    public void deleteSchedule(Long scheduleId, Long userId) {
+        Schedule schedule = getSchedule(scheduleId);
+        authValidator.validateAuthor(schedule.getUser().getUserId(), userId);
 
         scheduleRepository.delete(schedule);
+    }
+
+    private Schedule getSchedule(Long scheduleId) {
+        return scheduleRepository.findById(scheduleId).orElseThrow(
+                ScheduleNotFoundException::new
+        );
     }
 
 }
